+++ database/migrations/2022_07_05_083223_add_bahan_layak_dan_tidak_layak_to_lab_table.php
@@ -14,7 +14,7 @@ class AddBahanLayakDanTidakLayakToLabTable extends Migration
     public function up()
     {
         Schema::table('lab', function (Blueprint $table) {
-            $table->smallInteger('bahan_layak')
+            $table->float('bahan_layak')
                 ->nullable()
                 ->after('grid');
             $table->smallInteger('bahan_tidak_layak')
