+++ database/migrations/2022_07_25_021332_add_jumlah_to_detail_produksi_table.php
@@ -14,7 +14,7 @@ class AddJumlahToDetailProduksiTable extends Migration
     public function up()
     {
         Schema::table('detail_produksi', function (Blueprint $table) {
-            $table->integer('jumlah')
+            $table->float('jumlah')
             ->nullable()
             ->after('id_produksi');
         });
